# swagger-docs.json
{"openapi":"3.0.0","paths":{"/health":{"get":{"operationId":"HealthController_getHealth","summary":"","description":"Health check status","parameters":[],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseMessage"}}}}}},"400":{"description":"Invalid parameters","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseErrorMessage"}}}}}},"408":{"description":"Request timeout","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseErrorMessage"}}}}}},"422":{"description":"Bussiness rule","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseErrorMessage"}}}}}},"502":{"description":"Invalid response","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseErrorMessage"}}}}}}},"tags":["health"]}},"/services":{"post":{"operationId":"ServiceController_postService","summary":"","description":"Create service data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostServiceDto"}}}},"responses":{"201":{"description":"CREATED","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseMessage"}}}}}},"400":{"description":"Invalid parameters","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseErrorMessage"}}}}}},"422":{"description":"Bussiness rule","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseErrorMessage"}}}}}},"502":{"description":"Invalid response","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseErrorMessage"}}}}}}},"tags":["services"],"security":[{"ApiKeyAuth":[]}]},"get":{"operationId":"ServiceController_getServices","summary":"","description":"Get services","parameters":[{"name":"paginate","required":false,"in":"query","schema":{"default":true,"type":"boolean"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":50,"default":20,"type":"number"}},{"name":"name","required":false,"in":"query","description":"Name","schema":{"minLength":0,"maxLength":255,"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}},"400":{"description":"Invalid parameters","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseErrorMessage"}}}}}},"422":{"description":"Bussiness rule","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseErrorMessage"}}}}}},"502":{"description":"Invalid response","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseErrorMessage"}}}}}}},"tags":["services"],"security":[{"ApiKeyAuth":[]}]}},"/services/{id}":{"get":{"operationId":"ServiceController_getServiceById","summary":"","description":"Get service","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}},"400":{"description":"Invalid parameters","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseErrorMessage"}}}}}},"422":{"description":"Bussiness rule","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseErrorMessage"}}}}}},"502":{"description":"Invalid response","content":{"application/json":{"schema":{"properties":{"data":{"$ref":"#/components/schemas/ResponseErrorMessage"}}}}}}},"tags":["services"],"security":[{"ApiKeyAuth":[]}]}}},"info":{"title":"API NestJS with Oracle Database","description":"This application presents several programming concepts and best practices such as: \n    code organization and structure, design patterns, solid concepts and clean architecture.","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"ApiKeyAuth":{"type":"apiKey","in":"header","name":"x-api-key","description":"Security Token"}},"schemas":{"ResponseErrorMessage":{"type":"object","properties":{"message":{"type":"array","items":{"type":"string"}},"status":{"type":"number"},"statusText":{"type":"string"},"log":{"type":"string"},"path":{"type":"string"}},"required":["message","status","statusText","log","path"]},"ResponseMessage":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"PostServiceDto":{"type":"object","properties":{"name":{"type":"string","minLength":4,"maxLength":255,"description":"Name"},"description":{"type":"string","description":"Description"}},"required":["name"]}}},"externalDocs":{"description":"Organization Website","url":"https://example.com"}}